+++ R/emmGrid-methods.R
@@ -505,6 +505,10 @@ update.emmGrid = function(object, ..., silent = FALSE) {
 #'   \code{\link{summary.emmGrid}}, \code{\link{predict.emmGrid}}, \code{\link{test.emmGrid}},
 #'   \code{\link{confint.emmGrid}}, and \code{\link{emmip}}. The only option that can
 #'   affect the latter four is \code{"predict.method"}.}
+#' \item{\code{allow.na.levs}}{A logical value that if \code{TRUE} (the default), allows
+#'   \code{NA} to be among the levels of a factor. Older versions of \pkg{emmeans} did not
+#'   allow this. So if problems come up (say in a messy dataset that includes incomplete cases), 
+#'   try setting this to \code{FALSE}.}
 #' \item{\code{sep}}{A character value to use as a separator in labeling factor combinations.
 #'   Such labels are potentially used in several places such as \code{\link{contrast}} and 
 #'   \code{\link{plot.emmGrid}} when combinations of factors are compared or plotted.
@@ -680,6 +684,7 @@ emm_defaults = list (
     emmeans = list(infer = c(TRUE, FALSE)),
     contrast = list(infer = c(FALSE, TRUE)),
     save.ref_grid = FALSE,    # save new ref_grid in .Last.ref_grid
+    allow.na.levs = TRUE,     # Do we allow factors to have NA as a level?
     cov.keep = "2",           # default for cov.keep arg in ref_grid
     sep = " ",                # separator for combining factor levels
     parens = c("-|\\+|\\/|\\*", "(", ")"), # patterns for what/how to parenthesize in contrast